[INLONG-11560][Agent] Adjust the directory of data source extension plugins
